On deadfs return EIO error instead of EBADF
There are few places were EBADF is used and there are places where EIO
is used. Unify and use EIO everywhere as EIO matches this situation better.

Here is the error description:
"I/O error. This will happen for example when the process is in a
background process group, tries to read from its controlling tty, and
either it is ignoring or blocking SIGTTIN or its process group is
orphaned.  It may also occur when there is a low-level I/O error while
reading from a disk or tape."